+++ packages/react/src/get-client-browser-parameters.example.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import {useEffect} from 'react';
+import {getClientBrowserParameters} from '@shopify/storefront-kit-react';
+
+export default function App({Component, pageProps}) {
+  useEffect(() => {
+    console.log(getClientBrowserParameters());
+  });
+
+  return <Component {...pageProps} />;
+}
